fix formatting, tests

+++ srt-protocol/src/protocol/stats.rs
@@ -92,9 +92,9 @@ mod online_windowed_stats {
         assert_eq!(
             window,
             vec![
-                (ms(1000), 1000 as usize),
-                (ms(1000), 1000 as usize),
-                (ms(1000), 1000 as usize)
+                (ms(1000), 1000_usize),
+                (ms(1000), 1000_usize),
+                (ms(1000), 1000_usize)
             ]
         );
     }
